[WIP] drivers: i2c: stm32: SMBus in RTIO drivers
Add SMBus support in STM32 I2C RTIO drivers. For that purpose, factorizee
the function implemented in si2c_stm32.c (non-RTIO native driver)
to i2c_stm32_common.c.
